Renaming fingerprinting files to simplify and stop duplication in naming
